Refactored (and recoded) dashboard feature
Dashboard feature was previously poorly written, this commit reuses code
from requests feature to accomodate dashboard feature requirements.
Further review is needed.

+++ web-model/src/main/java/com/oasis/web_model/response/success/dashboard/DashboardStatusResponse.java
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class DashboardStatusResponse {
 
-    private int requestedRequests;
-    private int pendingHandoverRequests;
-    private int availableAsset;
+    private long requestedRequests;
+    private long pendingHandoverRequests;
+    private long availableAsset;
 
 }
